java: Move complex utility function to its own class
