+++ java/src/main/java/io/cucumber/cienvironment/DetectCiEnvironment.java
@@ -1,7 +1,5 @@
 package io.cucumber.cienvironment;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Map;
 import java.util.Optional;
 
@@ -17,13 +15,4 @@ public final class DetectCiEnvironment {
                 .findFirst();
     }
 
-    static String removeUserInfoFromUrl(String value) {
-        if (value == null) return null;
-        try {
-            URI uri = URI.create(value);
-            return new URI(uri.getScheme(), null, uri.getHost(), uri.getPort(), uri.getPath(), uri.getQuery(), uri.getFragment()).toASCIIString();
-        } catch (URISyntaxException | IllegalArgumentException e) {
-            return value;
-        }
-    }
 }
