+++ java/src/main/java/io/cucumber/cienvironment/RemoveUserInfo.java
@@ -1,0 +1,17 @@
+package io.cucumber.cienvironment;
+
+import java.net.URI;
+import java.net.URISyntaxException;
+
+final class RemoveUserInfo {
+    static String fromUrl(String value) {
+        if (value == null) return null;
+        try {
+            URI uri = URI.create(value);
+            return new URI(uri.getScheme(), null, uri.getHost(), uri.getPort(), uri.getPath(), uri.getQuery(), uri.getFragment()).toASCIIString();
+        } catch (URISyntaxException | IllegalArgumentException e) {
+            return value;
+        }
+    }
+
+}
